server/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-stubs==6.0.2
django-stubs-ext==6.0.2
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
gunicorn==25.3.0
packaging==26.0
psycopg==3.3.3
psycopg2-binary==2.9.11
psycopgbinary==0.0.1
sqlparse==0.5.5
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
